migrated existing dash routes to dash api route
