+++ packages/server/src/app/api/dashboard/dashboard.module.ts
@@ -1,15 +1,19 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 
-import { DashboardService } from './dashboard.service';
 import { NewsController } from './news/news.controller';
 import { NewsModule } from '../../db/news';
 import { getJwtSecretKey, JWT_EXPIRATION } from '../../util';
 import { UsersModule } from '../../db/users/users.module';
+import { ApprovalQueueModule } from '../../db/approval-queue/approval-queue.module';
+import { DocsModule } from '../../db/docs/docs.module';
+import { QueueController } from './queue/queue.controller';
+import { DocsController } from './docs/docs.controller';
+import { QueueService } from './queue/queue.service';
 
 @Module({
   imports: [
-    NewsModule, UsersModule,
+    NewsModule, UsersModule, ApprovalQueueModule, DocsModule,
     JwtModule.registerAsync({
       useFactory: () => ({
         secret: getJwtSecretKey(),
@@ -17,7 +21,7 @@ import { UsersModule } from '../../db/users/users.module';
       }),
     }),
   ],
-  controllers: [NewsController],
-  providers: [DashboardService]
+  controllers: [NewsController, QueueController, DocsController],
+  providers: [QueueService],
 })
 export class DashboardModule {}
