--- packages/server/src/app/api/dashboard/dashboard.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { DashboardService } from './dashboard.service';
-
-describe('DashboardService', () => {
-  let service: DashboardService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [DashboardService],
-    }).compile();
-
-    service = module.get<DashboardService>(DashboardService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
