--- packages/server/src/app/api/dashboard/dashboard.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class DashboardService {}
