+++ packages/server/src/app/api/dashboard/docs/docs.controller.ts
@@ -1,9 +1,8 @@
-import { Controller, UseGuards, Body, Param, Request, Get, Put, Patch } from '@nestjs/common';
-
-import * as models from '@pulp-fiction/models/docs';
+import { Body, Controller, Request, Get, Param, Patch, Put, UseGuards } from '@nestjs/common';
 import { Roles } from '@pulp-fiction/models/users';
-import { RolesGuard } from '../../../guards';
+import { CreateDoc, EditDoc } from '@pulp-fiction/models/docs';
 import { DocsService } from '../../../db/docs/docs.service';
+import { RolesGuard } from '../../../guards';
 
 @Controller('docs')
 export class DocsController {
@@ -11,7 +10,7 @@ export class DocsController {
 
     @UseGuards(RolesGuard([Roles.Admin]))
     @Put('create-doc')
-    async createDoc(@Request() req: any, @Body() docInfo: models.CreateDoc) {
+    async createDoc(@Request() req: any, @Body() docInfo: CreateDoc) {
         return this.docsService.createDoc(req.user, docInfo);
     }
 
@@ -34,7 +33,7 @@ export class DocsController {
 
     @UseGuards(RolesGuard([Roles.Contributor, Roles.Moderator, Roles.Admin]))
     @Patch('edit-doc')
-    async editDoc(@Request() req: any, @Body() docInfo: models.EditDoc) {
+    async editDoc(@Request() req: any, @Body() docInfo: EditDoc) {
         return await this.docsService.editDoc(req.user, docInfo);
     }
