+++ packages/server/src/app/api/dashboard/queue/queue.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ContribService } from './contrib.service';
+import { QueueService } from './queue.service';
 
-describe('ContribService', () => {
-  let service: ContribService;
+describe('QueueService', () => {
+  let service: QueueService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ContribService],
+      providers: [QueueService],
     }).compile();
 
-    service = module.get<ContribService>(ContribService);
+    service = module.get<QueueService>(QueueService);
   });
 
   it('should be defined', () => {
